♻️ Moved the email send pipeline out of the Email model
queueForSending, resumeSending, sendSingleRecipient and loadAttachments
move to EmailSendService in the api app. The moved code is byte-for-byte
identical apart from taking the email as a parameter instead of `this`.

This removes canSendFromEmail and getEmailBuilder from Email.ts, so only
the three EmailBuilder functions used by buildRecipients are left before
EmailBuilder itself can move out of @stamhoofd/models.

errorToSimpleErrors is shared by the pipeline and updateCount, so it moves
to its own helper module.

Email.test.ts moves along to EmailSendService.test.ts. Three loginDetails
assertions had to be corrected: $t is not initialised in the models test
environment, so both sides of the comparison returned the raw key and the
assertions passed without ever comparing the sentence. They now assert the
same html-escaped form the fourth test in that group already used.
