+++ backend/app/api/src/endpoints/global/email/PatchEmailEndpoint.ts
@@ -9,6 +9,7 @@ import { patchObject } from '@simonbackx/simple-encoding';
 import { SimpleError } from '@simonbackx/simple-errors';
 import { Context } from '../../../helpers/Context.js';
 import { EmailPreviewService } from '../../../services/EmailPreviewService.js';
+import { EmailSendService } from '../../../services/EmailSendService.js';
 
 type Params = { id: string };
 type Query = undefined;
@@ -209,7 +210,7 @@ export class PatchEmailEndpoint extends Endpoint<Params, Query, Body, ResponseBo
             }
 
             // Preview the sending status
-            await model.queueForSending();
+            await EmailSendService.queueForSending(model);
         }
 
         return new Response(await EmailPreviewService.getPreviewStructure(model, { allLanguages: true }));
